+++ rs/moq-transcode/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.4](https://github.com/moq-dev/moq/compare/moq-transcode-v0.0.3...moq-transcode-v0.0.4) - 2026-07-25
+
+### Other
+
+- *(moq-video)* [**breaking**] one raw Frame type, and carry timestamps through encode ([#2503](https://github.com/moq-dev/moq/pull/2503))
+
 ## [0.0.3](https://github.com/moq-dev/moq/compare/moq-transcode-v0.0.2...moq-transcode-v0.0.3) - 2026-07-24
 
 ### Other